requirements.txt
alembic==1.4.2
behave==1.2.6
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
flask-swagger-ui==3.25.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
parse==1.15.0
parse-type==0.5.2
public==2019.4.13
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
pytz==2020.1
query-string==2019.4.13
<<<<<<< HEAD
=======
requests==2.24.0
>>>>>>> develop
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==1.0.1
